requirements.txt
fastapi
uvicorn
python-dotenv
pytest
requests
httpx
tensorflow===2.13.0
dlib
opencv-python
numpy
pandas
pydantic
python-multipart